requirements.txt
jupyter
python-dotenv
tabulate
scipy